Moved active windows rendering into a dedicated component rather than just using the main App component.
